+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# TODO and Change Log
+# bract.ring TODO and Change Log
 
 ## TODO
 None
@@ -7,6 +7,13 @@ None
 ### Changed
 - Use `[bract/bract.core "0.3.0"]`
   - [TODO] Remove this when using final GA version
+- Add inducer `bract.ring.inducer/apply-wrappers` accepting Ring wrappers
+- [TODO - BREAKING CHANGE] Remove unparameterized wrapper-applying inducers in favor of `apply-wrappers`
+  - Remove inducer `bract.ring.inducer/ctx-apply-wrappers`
+  - Remove inducer `bract.ring.inducer/cfg-apply-wrappers`
+  - Remove config definition `bract.ring.config/ctx-wrappers`
+  - Remove config definition `bract.ring.config/cfg-wrappers`
+
 
 ## 0.2.0 / 2017-June-04
 ### Changed